responsible-ai-security-text-detector/requirements/requirement.txt
fastapi
uvicorn
pyyaml
transformers
tldextract
sentence-transformers
detoxify==0.5.0
pandas
python-dotenv
django-phonenumbers
pymongo
python-multipart
../lib/aicloudlibs-0.1.0-py3-none-any.whl
#../lib/en_core_web_lg-3.5.0-py3-none-any.whl
../lib/presidio_analyzer-4.0.6-py3-none-any.whl